+++ apim-portal/src/admin-portal/components/ManageConnectors/ViewConnector.tsx
@@ -135,8 +135,8 @@ export const ViewConnector: React.FC<IViewConnectorProps> = (props: IViewConnect
         <div className="p-ml-4">
           <div><b>Type</b>: {mo.apsConnector.connectorClientConfig.locationConfig.configType}</div>
           <div><b>Url</b>: {mo.composedConnectorUrl}</div>
-          <div><b>Service User</b>: {mo.apsConnector.connectorClientConfig.serviceUser}</div>
-          <div><b>Service Pwd </b>: {mo.apsConnector.connectorClientConfig.serviceUserPwd}</div>
+          {/* <div><b>Service User</b>: {mo.apsConnector.connectorClientConfig.serviceUser}</div> */}
+          {/* <div><b>Service Pwd </b>: {mo.apsConnector.connectorClientConfig.serviceUserPwd}</div> */}
         </div>
       </React.Fragment>
     );