+++ apim-server/server/common/ServerClient.ts
@@ -3,6 +3,8 @@ import { EServerStatusCodes, ServerLogger } from './ServerLogger';
 import { 
   OpenAPI 
 } from '../../src/@solace-iot-team/apim-server-openapi-node';
+import APSAuthStrategyService from './authstrategies/APSAuthStrategyService';
+import APSServiceAccountsService from '../api/services/apsAdministration/APSServiceAccountsService';
 
 
 export class ServerClient {
@@ -22,6 +24,10 @@ export class ServerClient {
     const base: URL = new URL(OpenAPI.BASE, `${ServerClient.protocol}://${ServerClient.host}:${ServerClient.expressServerConfig.port}${OpenAPI.BASE}`);
     // ServerLogger.trace(ServerLogger.createLogEntry(logName, { code: EServerStatusCodes.INFO, details: { base: base.toString() } } ));
     OpenAPI.BASE = base.toString();
+    // initialize with service account token for bootstrap calls
+    OpenAPI.WITH_CREDENTIALS = true;
+    OpenAPI.TOKEN = APSAuthStrategyService.generateServiceAccountBearerToken_For_InternalAuth({ serviceAccountId: APSServiceAccountsService.getInternalApsServiceAccountId() });
+
     ServerLogger.info(ServerLogger.createLogEntry(logName, { code: EServerStatusCodes.INITIALIZED, details: { OpenAPI: OpenAPI } } ));
   }